% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistar-configr.R
\docType{package}
\name{configr}
\alias{configr}
\alias{configr-package}
\alias{configr-package}
\title{R helpers for running data management workflows}
\description{
R helpers for running data management workflows
}
\details{
\tabular{ll}{
   Package: \tab configr\cr
   Type: \tab Package\cr
   Version
   : \tab 0.1\cr
   Date: \tab 2018-03-05\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
