% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistar.R
\docType{package}
\name{statistar}
\alias{statistar}
\alias{statistar-package}
\alias{statistar-package}
\title{R helpers for running data management workflows and enable R web computing}
\description{
R helpers for running data management workflows and enable R web computing
}
\details{
\tabular{ll}{
   Package: \tab statistar\cr
   Type: \tab Package\cr
   Version
   : \tab 0.1\cr
   Date: \tab 2018-03-05\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
